# Optimize `build.zig` discovery