Since #688 finding the `build.zig` for a given file has been made more accurate, but very inefficient.

The problems becomes very noticeable if you have a large project with many zig files.

## Benchmark
I've setup a small zig project that uses [zigwin32](https://github.com/marlersoft/zigwin32) with this main.zig:

```zig
const std = @import("std");
const win32 = @import("zigwin32/win32.zig");

pub fn main() !void {
    // put symbol here
}
```
Here are some measurements for requesting completions on the given symbols.
| Source | Build Mode  | Symbol           | Time |  Memory |
|--------|-------------|------------------|------|---------|
| Master | Debug       | win32.data       |  98s | 677.8MB |
| PR     | Debug       | win32.data       | <1ms |   1.2MB |
| Master | Debug       | win32.everything | 253s |   1.1GB |
| PR     | Debug       | win32.everything | 163s |   1.0GB |
| Master | ReleaseFast | win32.data       |  75s | 641.6MB |
| PR     | ReleaseFast | win32.data       | <1ms | 843.8kB |
| Master | ReleaseFast | win32.everything | 182s |   1.1GB |
| PR     | ReleaseFast | win32.everything | 127s | 956.0MB |